backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart>=0.0.6
openai==1.12.0
google-generativeai>=0.8.0
langchain==0.0.350
langchain-openai==0.0.2
chromadb>=0.4.22
# faiss-cpu>=1.9.0  # Optional - ChromaDB can work without it
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
beautifulsoup4==4.12.2
requests==2.31.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0
sentence-transformers>=2.3.0
huggingface_hub>=0.20.0
numpy>=1.24.3,<2.0.0
aiofiles>=23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0
httpx>=0.25.2
tenacity==8.2.3
python-dotenv==1.0.0
reportlab>=4.0.0
motor>=3.3.0
pymongo>=4.6.0
email-validator>=2.0.0
trafilatura>=1.6.0
lxml>=5.0.0
lxml_html_clean>=0.4.0
langdetect>=1.0.9
celery>=5.3.0
redis>=5.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
prometheus-client>=0.19.0
playwright>=1.40.0

